Merges both analytics modules
