+++ clients/cli/src/config.rs
@@ -40,30 +40,3 @@ impl Display for Environment {
         }
     }
 }
-
-mod analytics {
-    pub const DEV_MEASUREMENT_ID: &str = "G-SWNG3LZDFR";
-    pub const STAGING_MEASUREMENT_ID: &str = "G-T0M0Q3V6WN";
-    pub const BETA_MEASUREMENT_ID: &str = "G-GLH0GMEEFH";
-    pub const DEV_API_SECRET: &str = "8ySxiKrtT8a76zClqqO8IQ";
-    pub const STAGING_API_SECRET: &str = "OI7H53soRMSDWfJf1ittHQ";
-    pub const BETA_API_SECRET: &str = "3wxu8FjVSPqOlxSsZEnBOw";
-}
-
-pub fn analytics_id(environment: &Environment) -> String {
-    match environment {
-        Environment::Dev => analytics::DEV_MEASUREMENT_ID.to_string(),
-        Environment::Staging => analytics::STAGING_MEASUREMENT_ID.to_string(),
-        Environment::Beta => analytics::BETA_MEASUREMENT_ID.to_string(),
-        Environment::Local => String::new(),
-    }
-}
-
-pub fn analytics_api_key(environment: &Environment) -> String {
-    match environment {
-        Environment::Dev => analytics::DEV_API_SECRET.to_string(),
-        Environment::Staging => analytics::STAGING_API_SECRET.to_string(),
-        Environment::Beta => analytics::BETA_API_SECRET.to_string(),
-        Environment::Local => String::new(),
-    }
-}
